--- a/js/vm_cache.go
+++ b/js/vm_cache.go
@@ -14,10 +14,6 @@ const (
 
 	// cacheReapInterval is the amount to sleep in the cache reaper
 	cacheReapInterval = time.Minute
-
-	// cacheMaxConcurrency is the maximum number of concurrent accesses
-	// available for a given key.
-	cacheMaxConcurrency = 8
 )
 
 // vmCache provides an internal mechanism for caching javascript contexts
@@ -31,13 +27,7 @@ type vmCache struct {
 
 type cacheValue struct {
 	lastRead int64
-	vms      chan *otto.Otto
-}
-
-func newCacheValue() *cacheValue {
-	return &cacheValue{
-		vms: make(chan *otto.Otto, cacheMaxConcurrency),
-	}
+	vm       *otto.Otto
 }
 
 func newVMCache() *vmCache {
@@ -87,7 +77,10 @@ func (c *vmCache) Acquire(key string) *otto.Otto {
 	if !ok {
 		return nil
 	}
-	return <-val.vms
+	if val.vm == nil {
+		return nil
+	}
+	return val.vm.Copy()
 }
 
 // Init initializes the value in the cache, creating cacheMaxConcurrency copies
@@ -95,18 +88,6 @@ func (c *vmCache) Acquire(key string) *otto.Otto {
 func (c *vmCache) Init(key string, vm *otto.Otto) {
 	c.Lock()
 	defer c.Unlock()
-	val := newCacheValue()
+	val := &cacheValue{lastRead: time.Now().Unix(), vm: vm}
 	c.vms[key] = val
-	// Fill the cache with copies of the VM
-	for i := 0; i < cacheMaxConcurrency; i++ {
-		val.vms <- vm.Copy()
-	}
-	val.lastRead = time.Now().Unix()
-}
-
-// Release returns a VM to the cache.
-func (c *vmCache) Release(key string, vm *otto.Otto) {
-	c.Lock()
-	defer c.Unlock()
-	c.vms[key].vms <- vm
 }