… types/dynamic/redact.go
… types/dynamic/redact_test.go