This commit causes the javascript interpreter to always use a new
VM for every filter evaluation. This is much slower, but greatly
increases the safety of the feature, since users cannot cache state
between filter executions.

Before:
```
BenchmarkNewOttoVM-16   2572889 ns/op	 1514648 B/op	   13593 allocs/op
BenchmarkCheckEval-16     70336 ns/op	   18120 B/op	     405 allocs/op
BenchmarkMetricsEval-16   18662 ns/op	    4344 B/op	     120 allocs/op
```

After:
```
BenchmarkNewOttoVM-16   545758 ns/op	  321478 B/op	    3040 allocs/op
BenchmarkCheckEval-16   369857 ns/op	  189656 B/op	    1911 allocs/op
BenchmarkMetricsEval-16 315916 ns/op	  175872 B/op	    1626 allocs/op
```

The change dramatically increases CPU costs and GC pressure. It's
not known at this time if this can be improved. Profiles of the
Otto Copy() function haven't yielded any low hanging fruit for
optimization.

Entity redaction has been re-worked, and entities are now redacted before being filtered.

Signed-off-by: Eric Chlebek <eric@sensu.io>

Closes #2363 
Closes #2354