@echlebek 2018-11-19
@ccressent to address your question:

The VM that the user's code will run on is a copy of a VM that has had assets and utility functions preloaded. It is discarded after execution.